fix. rename and patch 1 out of 3 linter issues on front
